[feat] 5.1 가이드 업데이트

+++ manifest/2-1.install-capa.sh
@@ -1,3 +1,9 @@
+if [ ! -f "aws-credential.conf" ]; then
+    bash message.sh "ERROR" "'aws-credential.conf' is NOT EXIST!"
+    exit 0
+fi
+
+
 source aws-credential.conf
 
 cat << EOF | base64 | tr '\n' ' ' | sed 's/ //g' > tmp.txt
